+++ packages/fiori/src/NotificationListItemBase.ts
@@ -3,10 +3,8 @@ import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import { getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import { getTabbableElements } from "@ui5/webcomponents-base/dist/util/TabbableElements.js";
-import { isDesktop } from "@ui5/webcomponents-base/dist/Device.js";
 import getActiveElement from "@ui5/webcomponents-base/dist/util/getActiveElement.js";
 import ListItemBase from "@ui5/webcomponents/dist/ListItemBase.js";
-import Integer from "@ui5/webcomponents-base/dist/types/Integer.js";
 import { getEventMark } from "@ui5/webcomponents-base/dist/MarkedEvents.js";
 import { getFirstFocusableElement } from "@ui5/webcomponents-base/dist/util/FocusableElements.js";
 
@@ -31,7 +29,7 @@ class NotificationListItemBase extends ListItemBase {
 	 * @public
 	 */
 	@property()
-	titleText!: string;
+	titleText?: string;
 
 	/**
 	 * Defines if the `notification` is new or has been already read.
@@ -42,7 +40,7 @@ class NotificationListItemBase extends ListItemBase {
 	 * @public
 	 */
 	@property({ type: Boolean })
-	read!: boolean;
+	read = false;
 
 	/**
 	 * Defines if a busy indicator would be displayed over the item.
@@ -51,20 +49,20 @@ class NotificationListItemBase extends ListItemBase {
 	 * @since 1.0.0-rc.8
 	 */
 	@property({ type: Boolean })
-	loading!: boolean;
+	loading = false;
 
 	/**
 	 * Defines the delay in milliseconds, after which the busy indicator will show up for this component.
 	 * @default 1000
 	 * @public
 	 */
-	@property({ validator: Integer, defaultValue: 1000 })
-	loadingDelay!: number;
+	@property({ type: Number })
+	loadingDelay = 1000;
 
 	static i18nFioriBundle: I18nBundle;
 
 	get hasTitleText() {
-		return !!this.titleText.length;
+		return !!this.titleText?.length;
 	}
 
 	get loadingText() {
@@ -73,12 +71,6 @@ class NotificationListItemBase extends ListItemBase {
 
 	get isLoading() {
 		return this.loading;
-	}
-
-	onEnterDOM() {
-		if (isDesktop()) {
-			this.setAttribute("desktop", "");
-		}
 	}
 
 	/**