+++ packages/main/src/Tag.ts
@@ -17,14 +17,13 @@ import "@ui5/webcomponents-icons/dist/sys-enter-2.js";
 import "@ui5/webcomponents-icons/dist/error.js";
 import "@ui5/webcomponents-icons/dist/alert.js";
 import "@ui5/webcomponents-icons/dist/information.js";
-import WrappingType from "./types/WrappingType.js";
+import type WrappingType from "./types/WrappingType.js";
 import TagDesign from "./types/TagDesign.js";
-import TagSize from "./types/TagSize.js";
+import type TagSize from "./types/TagSize.js";
 // Template
 import TagTemplate from "./generated/templates/TagTemplate.lit.js";
 
 import {
-	TAG_DESCRIPTION_BADGE,
 	TAG_DESCRIPTION_TAG,
 	TAG_ROLE_DESCRIPTION,
 	TAG_ERROR,
@@ -84,8 +83,8 @@ class Tag extends UI5Element {
 	 * @public
 	 * @since 1.22.0
 	 */
-	@property({ defaultValue: TagDesign.Neutral })
-	design!: `${TagDesign}`;
+	@property()
+	design: `${TagDesign}` = "Neutral"
 
 	/**
 	 * Defines the color scheme of the component.
@@ -94,8 +93,8 @@ class Tag extends UI5Element {
 	 * @default "1"
 	 * @public
 	 */
-	@property({ defaultValue: "1" })
-	colorScheme!: string;
+	@property()
+	colorScheme = "1";
 
 	/**
 	 * Defines if the default state icon is shown.
@@ -104,31 +103,29 @@ class Tag extends UI5Element {
 	 * @since 1.22.0
 	 */
 	@property({ type: Boolean })
-	hideStateIcon!: boolean;
+	hideStateIcon = false;
 
 	/**
 	 * Defines if the component is interactive (focusable and pressable).
 	 *
-	 * **Note:** The tag cannot be `interactive`
-	 * when `design` property is `TagDesign.Set3`
 	 * @default false
 	 * @public
 	 * @since 1.22.0
 	 */
 	@property({ type: Boolean })
-	interactive!: boolean;
+	interactive = false;
 
 	/**
 	 * Defines how the text of a component will be displayed when there is not enough space.
 	 *
 	 * **Note:** For option "Normal" the text will wrap and the
 	 * words will not be broken based on hyphenation.
-	 * @default "None"
+	 * @default "Normal"
 	 * @public
 	 * @since 1.22.0
 	 */
-	@property({ type: WrappingType, defaultValue: WrappingType.None })
-	wrappingType!: `${WrappingType}`;
+	@property()
+	wrappingType: `${WrappingType}` = "Normal";
 
 	/**
 	 * Defines predefined size of the component.
@@ -136,29 +133,22 @@ class Tag extends UI5Element {
 	 * @public
 	 * @since 2.0.0
 	 */
-	@property({ type: TagSize, defaultValue: TagSize.S })
-	size!: `${TagSize}`;
+	@property()
+	size: `${TagSize}` = "S";
 
 	/**
 	 * Defines if the tag has an icon.
 	 * @private
 	 */
 	@property({ type: Boolean })
-	_hasIcon!: boolean;
+	_hasIcon = false;
 
 	/**
 	 * Defines if the tag has only an icon (and no text).
 	 * @private
 	 */
 	@property({ type: Boolean })
-	_iconOnly!: boolean;
-
-	/**
-	 * Defines if the tag has "Tag" design type.
-	 * @private
-	 */
-	@property({ type: Boolean })
-	_isTagDesign!: boolean;
+	_iconOnly = false;
 
 	/**
 	 * Defines the text of the component.
@@ -191,7 +181,6 @@ class Tag extends UI5Element {
 	onBeforeRendering() {
 		this._hasIcon = this.hasIcon || !!this._semanticIconName;
 		this._iconOnly = this.iconOnly;
-		this._isTagDesign = this.design !== TagDesign.Set3;
 	}
 
 	get _roleDescription() {
@@ -232,10 +221,6 @@ class Tag extends UI5Element {
 	get tagDescription() {
 		if (this.interactive) {
 			return undefined;
-		}
-
-		if (this.design === TagDesign.Set3) {
-			return Tag.i18nBundle.getText(TAG_DESCRIPTION_BADGE);
 		}
 
 		const valueState = this._valueState;