+++ packages/main/src/TimePickerInternals.ts
@@ -9,16 +9,12 @@ import getLocale from "@ui5/webcomponents-base/dist/locale/getLocale.js";
 import DateFormat from "@ui5/webcomponents-localization/dist/DateFormat.js";
 import getCachedLocaleDataInstance from "@ui5/webcomponents-localization/dist/getCachedLocaleDataInstance.js";
 import "@ui5/webcomponents-localization/dist/features/calendar/Gregorian.js"; // default calendar for bundling
-import CalendarType from "@ui5/webcomponents-base/dist/types/CalendarType.js";
+import type CalendarType from "@ui5/webcomponents-base/dist/types/CalendarType.js";
 import { fetchCldr } from "@ui5/webcomponents-base/dist/asset-registries/LocaleData.js";
-import Integer from "@ui5/webcomponents-base/dist/types/Integer.js";
 import UI5Date from "@ui5/webcomponents-localization/dist/dates/UI5Date.js";
-import SegmentedButton from "./SegmentedButton.js";
-import {
-	getHoursConfigByFormat,
-	getTimeControlsByFormat,
-	HourType,
-} from "./timepicker-utils/TimeSlider.js";
+import type SegmentedButton from "./SegmentedButton.js";
+import { getHoursConfigByFormat, getTimeControlsByFormat } from "./timepicker-utils/TimeSlider.js";
+import type { HourType } from "./timepicker-utils/TimeSlider.js";
 import {
 	TIMEPICKER_HOURS_LABEL,
 	TIMEPICKER_MINUTES_LABEL,
@@ -101,7 +97,7 @@ class TimePickerInternals extends UI5Element {
 	 * @default undefined
 	 * @public
 	 */
-	@property({ defaultValue: undefined })
+	@property()
 	value?: string;
 
 	/**
@@ -112,33 +108,33 @@ class TimePickerInternals extends UI5Element {
 	 * hh:mm:ss a -> 2:23:15 PM
 	 * mm:ss -> 12:04 (only minutes and seconds)
 
-	 * @default ""
+	 * @default undefined
 	 * @public
 	 */
 	@property()
-	formatPattern!: string;
+	formatPattern?: string;
 
 	/**
 	 * The index of the active Clock/TogleSpinButton.
 	 * @default 0
 	 * @private
 	 */
-	@property({ validator: Integer, defaultValue: 0, noAttribute: true })
-	_activeIndex!: number;
+	@property({ type: Number, noAttribute: true })
+	_activeIndex = 0
 
 	/**
 	 * Contains calendar type.
 	 * @private
 	 */
-	@property({ type: CalendarType })
-	_calendarType!: CalendarType;
+	@property()
+	_calendarType?: CalendarType;
 
 	/**
 	 * Contains currently available Time Picker components depending on time format.
 	 * @private
 	 */
-	@property({ type: Object, multiple: true })
-	_entities!: Array<TimePickerEntityProperties>;
+	@property({ type: Array })
+	_entities: Array<TimePickerEntityProperties> = [];
 
 	/**
 	 * Contains component-to-index map.
@@ -151,29 +147,29 @@ class TimePickerInternals extends UI5Element {
 	 * Contains currently available Button components depending on time format.
 	 * @private
 	 */
-	@property({ type: Object, multiple: true })
-	_periods!: Array<TimeSelectionPeriodProperties>;
+	@property({ type: Array })
+	_periods: Array<TimeSelectionPeriodProperties> = [];
 
 	/**
 	 * Id of the cooldown interval
 	 * @private
 	 */
-	@property({ validator: Integer, noAttribute: true })
+	@property({ type: Number, noAttribute: true })
 	_typeCooldownId?: ReturnType<typeof setTimeout>;
 
 	/**
 	 * Exact match number buffer
 	 * @private
 	 */
-	@property({ validator: Integer, noAttribute: true })
+	@property({ type: Number, noAttribute: true })
 	_exactMatch?: number;
 
 	/**
 	 * Buffer for entered by keyboard numbers
 	 * @private
 	 */
-	@property({ defaultValue: "", noAttribute: true })
-	_keyboardBuffer!: string;
+	@property({ noAttribute: true })
+	_keyboardBuffer = "";
 
 	static i18nBundle: I18nBundle;
 
@@ -278,7 +274,7 @@ class TimePickerInternals extends UI5Element {
 
 	get _formatPattern() {
 		const pattern = this.formatPattern;
-		const hasHours = !!pattern.match(/H/i);
+		const hasHours = !!pattern?.match(/H/i);
 		const fallback = !pattern || !hasHours;
 		const localeData = getCachedLocaleDataInstance(getLocale());