+++ packages/main/src/types/TableGrowingMode.ts
@@ -1,0 +1,20 @@
+/**
+ * Growing mode of the &lt;ui5-table&gt; component.
+ *
+ * @public
+ */
+enum TableGrowingMode {
+	/**
+	 * Renders a growing button, which can be pressed to load more data.
+	 * @public
+	 */
+	Button = "Button",
+
+	/**
+	 * Scroll to load more data.
+	 * @public
+	 */
+	Scroll = "Scroll",
+}
+
+export default TableGrowingMode;